requirements.txt
Flask==1.1.2
psycopg2-binary==2.8.6
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1